# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Тестовое задание",
        "contact": {},
        "version": "0.0.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/currency/add/{coin}": {
            "get": {
                "description": "adds currency for observing",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "currency"
                ],
                "summary": "adds currency for observing",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Coin Name",
                        "name": "coin",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.AddCurrency"
                        }
                    }
                }
            }
        },
        "/api/v1/currency/price/{coin}/{timestamp}": {
            "get": {
                "description": "get currency price by timestamp",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "currency"
                ],
                "summary": "get currency price",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Coin Name",
                        "name": "coin",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Timestamp",
                        "name": "timestamp",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.GetCurrencyPrice"
                        }
                    }
                }
            }
        },
        "/api/v1/currency/remove/{coin}": {
            "get": {
                "description": "get recommendations by user id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "currency"
                ],
                "summary": "delete currency from observing",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Coin Name",
                        "name": "coin",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.DeleteCurrency"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "response.AddCurrency": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
        "response.DeleteCurrency": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
        "response.GetCurrencyPrice": {
            "type": "object",
            "properties": {
                "coin": {
                    "type": "string"
                },
                "status": {
                    "type": "object",
                    "additionalProperties": true
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        }
    },
    "externalDocs": {
        "description": "OpenAPI",
        "url": "https://swagger.io/resources/open-api/"
    }
}